+++ src/avt/IVP/avtLCSIC.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2014, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -172,7 +172,7 @@ bool avtLCSIC::LessThan(const avtIntegralCurve *ic) const
 avtIntegralCurve*
 avtLCSIC::MergeIntegralCurveSequence(std::vector<avtIntegralCurve *> &v)
 {
-    for (int i = 0 ; i < v.size() ; i++)
+    for (size_t i = 0 ; i < v.size() ; i++)
     {
         avtLCSIC *ic = (avtLCSIC *) v[i];
