+++ src/avt/Math/avtBezierSegment.h
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2014, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -77,12 +77,12 @@ public:
 
     unsigned int degree() const 
     { 
-        return size()-1; 
+        return (unsigned int)size()-1; 
     }
 
     unsigned int order() const
     {
-        return size();
+        return (unsigned int)size();
     }
 
     avtVector firstV() const
