+++ src/avt/VisWindow/Tools/VisitSphereTool.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2014, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -708,7 +708,7 @@ VisitSphereTool::DoTransformations()
     if (proxy.Get3DAxisScalingFactors(axisscale))
         M = avtMatrix::CreateScale(axisscale[0],axisscale[1],axisscale[2]) * M;
 
-    for (int i=0; i<hotPoints.size(); i++)
+    for (size_t i=0; i<hotPoints.size(); i++)
         hotPoints[i].pt = M * origHotPoints[i].pt;
 
     vtkMatrix4x4 *tmp = vtkMatrix4x4::New();
