+++ src/databases/GULP/avtGULPFileFormat.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2014, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -778,7 +778,7 @@ avtGULPFileFormat::ReadTimestep(int ts)
 void
 avtGULPFileFormat::GetTimes(std::vector<double> &t)
 {
-    for (int i=0; i<times.size(); i++)
+    for (size_t i=0; i<times.size(); i++)
         t.push_back(times[i]);
 }
