+++ src/databases/MM5/avtMM5FileFormat.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2014, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -374,7 +374,7 @@ avtMM5FileFormat::GetMesh(int timestate, const char *meshname)
             // Default number of components for an array is 1.
             coords[i] = vtkFloatArray::New();
 
-            if (i < pos->second.size())
+            if ((size_t)i < pos->second.size())
             {
                 dims[i] = pos->second[i] + 1;
                 coords[i]->SetNumberOfTuples(dims[i]);
