+++ src/engine/main/DataNetwork.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2014, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -128,7 +128,7 @@ DataNetwork::~DataNetwork(void)
     if (!clone)
     {
         ReleaseData();
-        for (int i = 0 ; i < nodeList.size() ; i++)
+        for (size_t i = 0 ; i < nodeList.size() ; i++)
         {
             delete nodeList[i];
         }
