+++ src/engine/main/IceTNetworkManager.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2014, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -327,7 +327,7 @@ IceTNetworkManager::Render(bool checkThreshold, intVector networkIds, bool getZB
         this->RenderSetup(networkIds, getZBuffer, annotMode, windowID, leftEye);
 
         bool plotDoingTransparencyOutsideTransparencyActor = false;
-        for(int i = 0 ; i < networkIds.size() ; i++)
+        for(size_t i = 0 ; i < networkIds.size() ; i++)
         {
             workingNet = NULL;
             UseNetwork(networkIds[i]);
