+++ src/operators/ConnectedComponents/avtConnectedComponentsFilter.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2014, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -161,7 +161,7 @@ avtConnectedComponentsFilter::InferVariableNameFromContract(avtContract_p c)
     if (!foundIt)
     {
         std::vector<CharStrRef>   vars2nd = in_dr->GetSecondaryVariablesWithoutDuplicates();
-        for (int i = 0 ; i < vars2nd.size() ; i++)
+        for (size_t i = 0 ; i < vars2nd.size() ; i++)
         if (strncmp(*(vars2nd[i]), "operators/ConnectedComponents/",
                     strlen("operators/ConnectedComponents/")) == 0)
         {
