+++ src/operators/OnionPeel/OnionPeelViewerPluginInfo.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2014, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -254,7 +254,7 @@ OnionPeelViewerPluginInfo::InitializeOperatorAtts(AttributeSubject *atts,
             if (*collection != NULL)
             {
                 int nSets = collection->GetNumberOfSubsets();
-                for (size_t i = 0; i < nSets && !subsetNameValid; ++i)
+                for (int i = 0; i < nSets && !subsetNameValid; ++i)
                 {
                     bool isOn = false;
                     if (silAtts->GetTopSetIsAllOn())
