+++ src/operators/RemoveCells/QvisRemoveCellsWindow.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2014, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -222,7 +222,7 @@ QvisRemoveCellsWindow::UpdateWindow(bool doAll)
             const vector<int> &cV = atts->GetCellList();
             const vector<int> &dV = atts->GetDomainList();
             
-            for (int j = 0; j < cV.size(); ++j)
+            for (size_t j = 0; j < cV.size(); ++j)
                 cellList->addItem(CreateEntryString(cV[j],dV[j]));
             
             if (row > 0 && row < cellList->count())
